// tools/app-bootstrap/src/gqlTypes.d.ts
export type JSON = any;
// ...
  export type Variables = {
    switchesToLoad?: string[] | null;
  };

  export type Query = {
    __typename?: 'Query';
    dashboard?: Dashboard | null;
    rolesPermissionsData?: RolesPermissionsData | null;
  };

  export type Dashboard = {
    __typename?: 'Dashboard';
    id?: string | null;
    isConsoleUser?: boolean | null;
    user_id?: string | null;
    userIntercomHash?: string | null;
    sessionId?: string | null;
    employee?: Employee | null;
    company?: Company | null;
    switches?: JSON | null;
  };

  export type Employee = {
    __typename?: 'AllEmployee';
    id?: string | null;
  };

  export type Company = {
    __typename?: 'Company';
    id?: string | null;
  };

  export type RolesPermissionsData = {
    __typename?: 'RolesPermissionsData';
    grants?: JSON | null;
    spans?: JSON | null;
  };
// ...
export namespace WalkmeCompanySetupQuery {
  export type Variables = {};

  export type Query = {
    __typename?: 'Query';
    companySetupTasks: CompanySetupTasks[];
  };

  export type CompanySetupTasks = {
    __typename?: 'FlowSection';
    id?: string | null;
    isComplete?: boolean | null;
    name?: string | null;
  };
}
export namespace WalkmeUserDataQuery {
  export type Variables = {};

  export type Query = {
    __typename?: 'Query';
    dashboard?: Dashboard | null;
  };

  export type Dashboard = {
    __typename?: 'Dashboard';
    id?: string | null;
    isConsoleUser?: boolean | null;
    permission?: JSON | null;
    employee?: Employee | null;
  };

  export type Employee = {
    __typename?: 'AllEmployee';
    id?: string | null;
  };
}